Migrate stepper to styled-component
